+++ dcli/language/c.py
@@ -11,6 +11,16 @@ ASSET_PATHS = ['src/Makefile', 'src/template/Target.h', 'src/app/Logger.c', 'src
 
 TARGET_FILE = 'Target'
 
+BLANK = ' '
+
+TEMPLATE_SEPARATOR = ','
+
+CLOSE_TEMPLATE = '>'
+
+OPEN_TEMPLATE = '<'
+
+ARRAY_REPRESENTATION = '[]'
+
 
 class C(Language):
     def __init__(self):
@@ -25,3 +35,24 @@ class C(Language):
                           'c',
                           ASSET_PATHS
                           )
+
+    def format_data(self, datatype_holder):
+        formatted_data = ''
+        if datatype_holder.is_array:
+            formatted_data += datatype_holder.array_type.type_name
+            for i in range(0, datatype_holder.array_dim):
+                formatted_data += ARRAY_REPRESENTATION
+        elif len(datatype_holder.parameters_types) > 0:
+            formatted_data += datatype_holder.parametrized_root_type.type_name + OPEN_TEMPLATE
+            param_length = len(datatype_holder.parameters_types)
+            for i in range(0, param_length):
+                if param_length > 1 and i > 0:
+                    formatted_data += TEMPLATE_SEPARATOR + datatype_holder.parameters_types[i].type_name
+                else:
+                    formatted_data += datatype_holder.parameters_types[i].type_name
+            formatted_data += CLOSE_TEMPLATE
+        else:
+            formatted_data += datatype_holder.type_name
+        if datatype_holder.is_arg:
+            formatted_data += BLANK + datatype_holder.arg_name
+        return formatted_data