+++ dcli/language/cpp.py
@@ -37,25 +37,25 @@ class Cpp(Language):
                           )
 
     def format_data(self, datatype_holder):
-        ret = ''
+        formatted_data = ''
         if datatype_holder.is_array:
-            ret += datatype_holder.array_type.type_name
+            formatted_data += datatype_holder.array_type.type_name
             for i in range(0, datatype_holder.array_dim):
-                ret += ARRAY_REPRESENTATION
+                formatted_data += ARRAY_REPRESENTATION
         elif len(datatype_holder.parameters_types) > 0:
-            ret += datatype_holder.parametrized_root_type.type_name + OPEN_TEMPLATE
+            formatted_data += datatype_holder.parametrized_root_type.type_name + OPEN_TEMPLATE
             param_length = len(datatype_holder.parameters_types)
             for i in range(0, param_length):
                 if param_length > 1 and i > 0:
-                    ret += TEMPLATE_SEPARATOR + datatype_holder.parameters_types[i].type_name
+                    formatted_data += TEMPLATE_SEPARATOR + datatype_holder.parameters_types[i].type_name
                 else:
-                    ret += datatype_holder.parameters_types[i].type_name
-            ret += CLOSE_TEMPLATE
+                    formatted_data += datatype_holder.parameters_types[i].type_name
+            formatted_data += CLOSE_TEMPLATE
         else:
-            ret += datatype_holder.type_name
+            formatted_data += datatype_holder.type_name
         if datatype_holder.is_arg:
-            ret += BLANK + datatype_holder.arg_name
-        return ret
+            formatted_data += BLANK + datatype_holder.arg_name
+        return formatted_data
 
     def format_generic_declaration(self, type_name):
         return 'typename ' + type_name