+++ dcli/template/python/src/main/python/success/__init__.py
@@ -1,5 +1,7 @@
 def {{targetMethod}
 
+}
+
 }({{targetMethodArgs}}):
 # implement your own solution
 pass