+++ dcli/template/python/src/main/python/template/__init__.py
@@ -1,4 +1,6 @@
 def {{targetMethod}
 
+}
+
 }({{targetMethodArgs}}):
 pass