+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## UNRELEASED
 * fix sping error when build fail
 * create docker env to run|solve metamorph challenge instead of using host binary
+* better logs when error happens to run metamorph mission
 
 ## 0.0.98 (27/02/2020)
 * fix missing pyyaml dependency