+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 
 ## UNRELEASED
+
+## 0.0.99 (09/03/2020)
 * fix sping error when build fail
 * create docker env to run|solve metamorph challenge instead of using host binary
 * better logs when error happens to run metamorph mission